requirements.txt
mysql-connector-python
pinecone-client
regex
sentence-transformers
transformers
torch
